+++ src/com/tw/bootcamp/problem1/model/Rectangle.java
@@ -2,29 +2,29 @@ package com.tw.bootcamp.problem1.model;
 
 public class Rectangle implements Polygon {
 
-    private final int width;
-    private final int length;
+    private final double width;
+    private final double length;
 
-    private Rectangle(int width, int length) {
+    private Rectangle(double width, double length) {
         this.width = width;
         this.length = length;
     }
 
-    public static Rectangle createRectangle(int width, int length) {
+    public static Rectangle createRectangle(double width, double length) {
         return new Rectangle(width, length);
     }
 
-    public static Rectangle createSquare(int side) {
+    public static Rectangle createSquare(double side) {
         return createRectangle(side, side);
     }
 
     @Override
-    public int calculateArea() {
+    public double calculateArea() {
         return length * width;
     }
 
     @Override
-    public int calculatePerimeter() {
+    public double calculatePerimeter() {
         return 2 * (length + width);
     }
 }
